task_management/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-extensions==3.2.3
django-pytest==0.2.0
djangorestframework==3.15.2
flake8==7.1.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.2
pytest-django==4.9.0
sqlparse==0.5.1
